src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
PyMuPDF==1.24.3
motor==3.6.0
pymongo==4.9.0
pydantic-mongo==2.3.0
openai==1.35.13
cohere==4.4.1
google-genai==1.8.0
qdrant-client==1.10.1
groqcloud==0.4.0
mistralai==1.7.0
python-jose==3.3.0
supabase==1.3.3
